Add a predicate for FairServerMessageQueue which will attempt to check if a send would be successful before selecting the given message as the front item on the queue.  This can be guaranteed for SST, but not for RakNet, so we still do a check that the send was indeed successful.
Track which queue we've said holds the front message in FairQueue to ensure we use that queue in pop().  This was a problem because the network was being serviced in both of these methods previously, so a message that previously would have been vetoed might get through.  Now we only do that servicing in front(), so the results from that might be inconsistent, but pop() will always pop the last message returned by front().