Adding a SpaceProxyManager so that the space may communicate to any proxy objects that its plugins may request This necessitated giving each space node (soon Space::Space will be renamed to Space::Node or Space::Server ) its own Object ID unique to that space node This ObjectID would not be advertised to users, but it would be used to make requests (eg proximity, object property requests)
The one wrinkle is that some system needs to track the Proximity query needed for the ghost regions around physics.  I called this subsystem Space::Physics and it will also have to be in charge of asking for the appropriate object properties to make physics go...
This functionality will be shared with HostedObject, so it might make sense to have a common base class.

Further discussion is needed of course