Be more restrictive about when we say Mono was found.
Check that all the runnables were actually found instead of just
setting the variables and ignoring whether valid values were actually
placed in them. Ideally this would work in the opposite way where we
mark as not found by default and then set the variable on if
requirements are satisfied, which is the normal pattern for CMake
scripts.