Use an ODP::Port to handle RPC messages in HostedObject.
The original branch point for RPC messages is still there but marked
as deprecated (and tried after the ODP::Port rout).  This is currently
necessary because of bootstrapping -- the registration messages need
to be delivered before ODP::Ports can be allocated.  Eventually this
should be the responsibility of the object host.