Implement ODP::Service interface for SpaceProxyManager via DelegateService and sending via the parent Space.
Currently this is broken in the sense that in order to send we need to
use the parent Space's processMessage() method, which means we can't
do sanity checking to ensure a connection to the object exists and we
don't get feedback as to whether or not the message was accepted.  In
the long term, either we need to expose this return value or
SpaceProxyManager should not implement this interface but Space
should.  Since its actually the Space that acts as the object, the
latter is probably the right solution.