Add property access to visible data.
Convert Visible to a FunctionTemplate, expose it so we can add to its
prototype, and add property accessors for the usual attributes so they
don't need to be accessed via methods. Hide the constructor in the
shim layer and also add pretty print overrides to print the standard
set of properties.