'Fix' models using textures for transparency by ignoring them.
So far we've only encountered on model using this feature and it
rendered incorrectly because we turned on the blending for
transparency but then never provided correct data to get transparency
values. It turns out the model is likely incorrectly specifying this,
so for now we'll just ignore transparency textures. The real solution
probably requires adding a pass to apply the transparency first, then
doing another pass for normal shading. Fixes #219.