Fix SessionManager bug where invoking callbacks in loop could cause crash.
Make SessionManager copy set of disconnected objects before starting
to invoke their callbacks. Previously it was looping over a set of
objects and invoking each callback, but the callbacks could (and did)
make requests back to the SessionManager which affected the data being
looped over. Making a copy ensures all the callbacks are invoked.
Fixes #255 since the notification was already being made, but was
being missed because the client ended up crashing before it could be
seen.