Make avatar controllable by itself via Movable interface.
Make Movable take an optional boolean to control whether messages from
others are filtered out. This allow a movable to use only that
interface but restrict access. Add movable interface back to avatar,
but with this restriction. This allows interactions that work on other
objects, e.g. setting physics parameters, to also work on your own
avatar as long as you can select it, i.e. if you are in 3rd person mode.