Make Storage a standard OH component and provide file-based plugin as default.
Move the previously named JSBackendInterface to liboh as
Sirikata::OH::Storage and load an instance as a standard part of the
object host, based on a few new settings (oh.storage and
oh.storage.opts). Move the file based implementation into an oh-file
plugin and make it the default. Remove the createEntry API from the
class, instead having it autogenerate them on writes.