Make script text a standard, fixed argument to ObjectScript creation.
Make all ObjectScriptManagers use a standard script argument for
accepting the initial script. Originally we used an import file
argument, but as we've extended this to be generic scripts (with just
an import statement in most cases), it has become difficult to
shoehorn the script into the standard arguments list. Instead,
everywhere that used to just use script type and args now also has a
script contents argument.

This requires changing not only the ObjectScriptManager interface, but
also the object factories, persisted object set, and the
implementation of a few Emerson script methods.