Fix exceptions thrown by initial script execution.
If the initial code executed by a script threw an exception, we were
getting v8 into a bad state, causing a crash as soon as any other
object tried to use v8.  This happened because we were throwing
exceptions in C++ but no v8 code was there to catch it -- a
v8::TryCatch only catches exceptions if they pass through JS code
between the thrower and the TryCatch.  To fix this, make the basic
utilities for evaluating code -- internalEval and protectedEval --
optionally return the exception (discarding the normal return value)
instead of throwing them.  The default is still to throw the
exceptions and this is only used by the initialization code right now.
Fixes #271.