Add SQLiteStorageTest.
Add a StorageTestBase which provides all the real testing of the
OH::Storage interface. Use it to implement SQLiteStorageTest. This
also required adding the ability to run a Context only in separate
threads, allowing the initiating thread to continue, since the test
runner can't handle entering an event loop. Fixes #277.