Add very generic raytrace method and more commonly useful overloads.
Add a std.raytrace.raytrace method which is the most general, allowing
control over input and output coordinate frames, which mesh to test,
and the ray to check. Provide overloads raytrace (for world
coordinates only), raytraceFrom (for starting the ray in this's
coordinate frame) and raytraceTo (for testing against this's mesh) to
visibles and presences.

Fixes #310.