Handle negative local-time update times.
Sometimes the server will return update times that result in a
negative local time, but since Time uses a uint64 it results in never
being able to update ProxyObjects because the times are always smaller
than the negative number interpreted as a positive one. Instead, clamp
times to 0 if they will underflow. A better solution would be to
adjust these values, but this works fine for now. Fixes problem where
restored objects would respond to other property changes but not to
location or orientation updates.