Make sure restored presences get associated with their ProxyObject.
Because restored presences are created before connection is complete,
they miss the creation of their ProxyObject and setting up a
ProxyCreationListener doesn't work because the ProxyObject is already
created by the time the subscription to the ProxyManager is
setup. Instead, when we setup the ProxyCreationListener, force
notification about the self ProxyObject, thereby forcing the listeners
for mesh and position changes to be setup as well. Fixes #342.