Setup default.em to store basic presence data every 60 seconds.
This isn't a long term solution, but its a good choice for static
content, ensuring we get a sort of snapshot, and covers the majority
of objects (which use only one presence). It can also be overridden,
so making the default isn't problematic if you create an object with
it and then want to disable it.