Name meshes/billboards in Ogre by a hash of all data.
Compute a hash over all the data for a mesh (hash of data and
dependent assets) and use that for the name in Ogre. This is necessary
because the original asset (e.g. the mesh) may have an identical hash
to another, but use names in it to refer to dependent data
(e.g. textures) which, when turned to absolute URLs, are not the same
and point to different data.