Move URI class to URL, make URIs generic.
The old URI class was actually for URLs, so we move it to URL and make
the URI class just decode the scheme. All the transfer code still uses
URIs and dispatches based only on the scheme, but where necessary we
can compute the URL -- mostly used for computing new relative URLs
based on an old one.