Fix restoration from storage.
Fix restoration of presences from storage. First, add an intermediate
phase for objects in nameService, allowing them to be known, but still
unavailable. Change restore.em to use this intermediate state to avoid
processing objects twice, which was leading to multiple loads and
callbacks for the same object. By fixing this, we also change a
condition which, due to the switch to asynchronous loading, was being
triggered as true since the load hadn't finished and the name service
still had a DNE, resulting in incorrect callbacks indicating failure.