Provide different SessionEventListener callbacks for connection and migration.
Add SessionEventListener::onMigrated callback and trigger it instead
of SessionEventListener::onConnected after a stream is finished
connecting after a migration. This was necessary because the
notification from HostedObject comes after stream setup to ensure all
the components users (scripts) might need are available before the
callback. This requires changing a few callback signatures and
carrying some state through some callbacks in SessionManager since at
that point it only knew it was connected, not how it had gotten
connected.

Fixes #367 because the onConnected callback which was triggering an
attempt to double allocate an ODP port is no longer triggered.