Fix unsafe multithreaded access to Proximity stream info.
When sequence numbers were added to Proximity, it was added to the
ProxStreamInfo struct. This ended up not being safe because the map
storing ProxStreamInfo structs could only be safely accessed by the
main thread, but the sequence numbers needed to be used when
generating result messages, which is performed in the prox thread.

Instead of combining all the data, this patch stores the sequence
number information in a separate map which is only accessed by the
prox thread. Sequence numbers are also used by Loc, but the data that
needs to be passed is thread-safe -- a shared pointer to an atomic
int.