Check for existing presences when restoring a presence.
When restoring a presence, check if it already exists in the presences
map before trying to create and connect it. This case occurs when you
reset and the reset script tries to load from storage, including
restoring presences (i.e. if you use reset with the current version of
default.em). Fixes #349.