Fix inter-space Proximity messages.
When sequence numbers were added for Loc and Prox updates, they were
marked required but weren't added to server messages. Add them and the
necessary tracking of SeqNoPtrs to Proximity. Because bogus sequence
numbers were used for location messages, we only needed to swap in the
correct SeqNoPtr to fix them.