Make SingleMaterialGeometryFilter use much less memory.
Rework SingleMaterialGeometryFilter to reduce the vertex data as it
copies each SubMeshGeometry for each material used. The only case it
should still use a lot of memory is if there is a large
SubMeshGeometry containing primitives using different materials but
which share a lot of the underlying vertex data (therefore requiring a
lot of duplication). Fixes #401.