Apply scale-to-unit-size transformation before raytracing.
While meshBounds was taking into account the scale-to-unit-size
transform that needs to be applied to meshes before using them, the
raytrace function was not. For meshes that weren't even close to unit
scale, this resulted in odd results, often way outside the bounds of
the object. Compute the scale factor and apply it by adding a
transformation parameter to the raytracing methods. Fixes #406.