Centralize mesh loading decisions and add limit to maximum simultaneous meshes.
Rename a few methods in Entity and centralize mesh loading decisions
in the DownloadPlanner classes. Add a cap to the maximum number of
objects and besides loading, make the DownloadPlanners track whether
they need to swap currently loaded and unloaded objects due to changes
in priority. Reduce the frequency of these checks as we generally
don't expect such fast changes. Fixes #404.