Make Ogre get closer to target frame rate.
Add a flag, off by default, which makes Pollers (-> PollingService ->
TimeSteppedSimulation -> TimeSteppedQueryableSimulation ->
OgreRenderer) optionally use a more accurate mode which measures the
time spent in the user's callback and accounts for it when setting the
next timeout. This still gracefully handles falling behind and is
still approximate, but accounts for the case where the user callback
takes a long time compared to the frequency of callbacks requested. So
far this is only enabled for Ogre since other Pollers don't care about
the accuracy and in some cases the performance penalty would be too
high.