Use data identifiers, not unique identifiers, for TransferRequests.
Previously TransferRequests were trying to generate unique
identifiers. This was both broken (sometimes on Windows you would get
duplicate identifiers) and defeated the AggregateRequest design since
requests couldn't be properly identified as identical.  The variable
is renamed so it doesn't indicate uniqueness and the proper identifier
is used for each type (URL or hash).  This allows proper aggregation
of requests from different TransferPools.