Fix method deciding whether a prox query handler should track an object.
Fix a missing check for local vs. replica objects when deciding
whether to handle objects in LibproxProximity. Because this was
missing, both the local (server-to-server query) and global
(object-to-server) query handlers would track all objects the server
heard about. This probably had a number of negative effects (lots of
extra messaging, extra effort tracking these objects and memory usage
in the query handler data structures, etc), but in particular it
resulted in a ping-pong effect between servers when using rtreecut
handlers. Additions and removals caused cuts to be lifted, resulting
in the other server also adjusting its tree and triggering cut lifts,
and so on, back and forth between servers.  Ultimately this made it to
the client, resulting in flickering objects.