Replace query parameters with opaque query string.
Update sirikata-protocol to a version with opaque query strings and
which deprecates the solid angle and max count parameters. Convert
ObjectHost and HostedObject to use these, supporting translation of
the old parameters. Deprecate or remove OH functions that used the old
format. Update the space server to use the new query string by
default, but fall back to the old format to support old clients.