Make IOTimer have stranded and non-stranded modes, use stranded mode in MigrationMonitor.
Previously, IOTimer tried to work around (and check with
SerializationCheck) the multithreaded event queue, hoping to force
access to a single thread. However, because the timer callback it
registered could happen on any thread, it isn't possible to safely
implement this without locks. Instead, this modifies IOTimer to
optionally use an IOStrand. The IOStrand provides serialization and
additional logic takes care of not invoking callbacks if they were
triggered (but not handled) before cancel() was called.

This also converts MigrationMonitor to use this stranded mode since it
was intended to operate within a single strand anyway and I observed
crashes due to it when the space server was under moderate load.