Add refine/coarsen requests to manual query processor.
Add encoding and sending of refine and coarsen requests to the manual
object query processor. Add some test code which forces refinement on
any additions which are known to be aggregates. This is only a
temporary solution but will allow testing of code on the server and
client without a more complicated approach.