Parse refine/coarsen requests and forward them to queries.
Parse refine and coarsen requests. Forward them to queries, which in
turn pass them to the query handler for processing. This finally
allows clients to control their cuts.  Also update libprox for a fix
to refinement handling where leaf nodes were not refined due to a
logic error: leaves can be refined (to get at individual objects) so
long as they are still in the result set.