Add and use LocationServiceCache for OH queries.
Add a LocationServiceCache for OH queries which uses
SequencedPresenceProperties for its underlying storage. This version
is thread-safe but guarantees Prox LocationUpdateListener events are
triggered in a strand. Replace the old storage in
ManualObjectQueryProcessor with this one.

This change also required a number of other semi-related changes. The
interface for libprox LocationServiceCache needed to be updated to
remove consts to allow locks and return-by-reference to allow
returning values derived from other values. Also reorganized the
ProxSimulationTraits from a monolithic version in libspace to
situation-specific versions -- they share some overlap, but are small
and this gets rid of awkward sharing in a library which didn't need
those types. Also add a few ObjectReference nested types for
generating random and null ObjectReferences.