Add character support to bullet physics location service.
Add BulletCharacterObject as a new subclass of BulletObject and add
the necessary options to physics settings to instantiate it. It uses a
slightly modified version of Bullet's own
btKinematicCharacterController, which uses a ghost object to detect
and respond to collisions in a non-physical way. This also changes the
default scene to include physics settings so it will act as a simple
test environment, especially since it has a terrain to test character
walking with.