Fix runaway forked processes.
cppoh was sometimes leaving runaway processes eating up CPU upon
exit. It turns out this is because the forked process was trying to
cleanup SDL and getting itself into a bad state. For now, just don't
call SDL_Quit when cleaning up SDL, which isn't great, but individual
components can be cleaned up by the users of libsdl anyway. Also, SDL
is currently only used by clients, which generally tie the lifetime of
SDL to the lifetime of the process anyway.