Fix repeated use of IOTimer after cancel.
Remove an unnecessary reset of mCanceled in IOTimer which was causing
the comparison of the value in the callback to be false if the timer
had previously been canceled. The reset was unnecessary and looked
like it was leftover from a previous version of the class since it was
using a boolean where we actually use a uint32.