Fix masking of object removals in graphics.
Move control of adding and removing objects from the download planner
to ProxyEntity, which has the timer that masks temporary removal of
objects from the result set. Add a boolean to track whether the
ProxyEntity is currently active in the download planner and use this
to decide whether to remove it when the timeout is invoked.