Use ogre.cfg from hidden user-specific directory by default.
Make the default offset for ogre.cfg the user-specific hidden
directory, which makes the settings global for the user which is more
convenient. Also add a note that we specifically don't do this for
Ogre.log, which ends up relative to the directory being run from,
thereby separating logs and keeping them organized by where they were
generated from.