Add path utility for finding shared resources.
Shared resources are files which would get installed under share/, but
which may be scattered throughout the source tree. To support this,
instead of looking up a single path, the resources are found using a
resource file/directory to search for and an in-tree alternative
path. In both cases we can search starting from the binary's location
and check only a few (sometimes platform-dependent) options. Also add
support for a placeholder and use this in the JS plugin to provide a
single default import path.