Fix bug where setup of query wasn't necessarily complete when callbacks from libprox could start.
Previously some information about queries was being setup after the
Proximity class was setup as the event listener. Since a callback can
be triggered in the event listener setter when events are available,
this could result in incorrect handling of these results. Instead,
this just reorders the setup to make sure the information needed is
already available if the callback is made immediately. It also adds a
couple of assertions to make sure we have the information available in
the callback.