Fix possible 'infinite loop' in LibproxProximityBase.
In the case where a base SST stream wasn't available yet, we would
just re-post, periodically retrying. This is usually fine, but if an
object connects, registers a query, and disconnects before a stream is
created then we can get stuck retrying indefinitely. Instead, check
for valid sessions on each iteration, which is always true when the
object is still connected, even if the base stream is not there yet.