Remove predictive updates for ProxyObjects.
Remove predictive updates from ProxyObjects (and
SequencedPresenceProperties). ProxyObjects will always reflect the
most up-to-date *confirmed* view of objects in the world. Since we
only need predictions on presences, we can maintain them only for
presences; in fact, we already do since we need to track the values in
order to send out requests.