Add predicted/verified location accessors and use predicted by default.
Provide two sets of accessors in ProxyObject: one for
predicted/requested location and one for verified. Most ProxyObjects
just use the verified data because they are not presences. For
presences they try to look up the data and use the epoch information
to decide whether to use the requested information or the data stored
in the ProxyObject (i.e. data received from the space that hasn't been
overwritten with predicted data). Remove some accessors which used
data directly and fix up extrapolateLocation to use the predicted
values, which seems to be what all uses of it expected.

Along with previous few commits, fixes #472, which addressed this
particular issue, and fixes #467, which was a result of not resolving
conflicting requested and reported location information properly.