Avoid very small updates to characters in Bullet LocationService.
In the Bullet LocationService, because characters need to be handled
in a custom way, we don't get the benefits of objects settling down
like other Bullet objects. Instead, detect that minimal change has
occurred and skip the update when possible. This avoids constantly
sending updates when nothing is really happening. Fixes #471.