Properly handle location update requests for physics objects.
In Bullet plugin, handle location updates by applying updates directly
if physics is off or the object is remote and by deferring to the
BulletObject to (possibly) apply the update otherwise. This allows
objects to be moved even when they are being simulated by Bullet. For
requests from objects, the BulletObject can decline the
update. Currently this is only used for static objects, which are set
that way because they *shouldn't* be moving anyway. This ensures that,
e.g., a ground plane does not move. Dynamic objects just apply updates
and continue, so this fixes #258. It can cause things to pop (or push
objects beneath the ground!), but nothing worse than what would happen
if you turned physics off, moved the object, and then turned physics
on again.