Fix use of invalid reference into a vector.
The collada parser was saving a reference to a field in the last
element of a vector, but during the loop where it was used we possibly
added another element to that vector, invalidating the reference. Just
lookup the field everytime since we'd need to change the
pointer/reference when we push the new element anyway. Fixes #477.