Update proximityAdded unit test for proximity results that don't include self.
The proximityAdded unit test checks for the number of visibles
found. Recently, 5d366a822ef2fb57cd9f64cc4f6085c0a635fdfa made it so
a visible for the querying presence is not returned. This causes one
fewer objects in the result set. Adjust the test and notes
accordingly.