Add meerkat:// upload support.
Add upload support for the meerkat:// protocol. This add
UploadRequests and makes them work through the
TransferMediator. Uploads always have a few required parameters, file
data, and an optional set of parameters. When handling the upload, one
implementation for meerkat:// URLs, implemented in
MeerkatUploadHandler, handles the actual OAuth upload to the CDN,
checks the status URL until it finishes or times out, and returns the
new URL.