Make use of sequence numbers to better handle retries.
Use session seqnos in Server to better handle duplicate/different
requests and use tracked seqnos in SessionManager to reuse the same
request ID for retries and clear the recorded seqno when replies come
back. Note that seqno is actually a bit of a misnomer -- they are
really more like session identifiers.

The basic approach is to start a with a new seqno in SessionManager
for each connection request. This allows both sides to handle
sequences of session messages triggered by an original request
separately. Both sides check/update sequence numbers and which node
they are talking to (ServerID or OH Connection ID) for each
request. Mismatching values usually result in the node ignoring the
message, except in the few cases where doing so could trigger getting
into a bad state (in particular, an unexpected successful connection
response requires the OH to try to kill that connection). Matching
values are handled as they previously were.

Fixes #482.