Fix bugs in Emerson's handling of query settings.
The shift to using a single string for the query settings introduced
some bugs: an incorrect check for a number instead of a string,
failure to handle some edge cases (e.g. empty string returned by
query, which causes JSON.parse to fail), and wrappers such as
setQueryAngle which didn't pass the right value back to
setQueryAngle.