Make Emerson Presences get query from parent HostedObject when connected.
The management of queries is a bit tricky because it isn't handled by
the regular properties tracked for each visible. We need separate
storage because we need a place to store the data when restoring a
presence. However, when we get connected, we need to use the value
stored in the HostedObject's PerPresenceData. This change just chooses
where to look for the data based on whether the object is
connected. Fixes #488.