Fix some cases where OHLocationServiceCache didn't track objects for long enough.
Add a simple kind of tracking that just increases/decreases the
refcount without returning an iterator -- the object properties still
need to be looked up by object identifier. Use this to cover periods
where we want to check for an object and check some of its properties
and be certain the object won't be removed between those operations.

The previous change also exposed an issue where if tracking stopped
while the object was still part of a result set we could lose it in
the cache. Add refcount increment/decrement on object addition/removal
so we are ensured the object won't be removed until it has both left
the result set and has no external tracking calls outstanding.