Fix off-by-one error that caused Contexts to run one fewer thread than requested.
When the ability to allocate all new threads was introduced, an
off-by-one error was added which caused one fewer thread than
requested to be allocated (in both all-new and use-original-thread
cases).

As a result, the increase in threads on the object host didn't have
the intended effect. There are definitely still issues with
multithreading, so this reverts that (previously ineffective) change.

Also remove a thread ID check from HostedObject, which shouldn't be
necessary and throws an assertion -- we want HostedObjects to be able
to run from different threads now. Finally, because this was the only
use of the ThreadId class, which essentially just wrapped boost's
thread::id, remove ThreadId entirely.