Improve COLLADA export support for materials.
Pass all types of effects except opacity through to the COLLADA
encoder (diffuse, specular, ambient, emission, and
reflective). Instead of assuming a Phong material, determine which
type to emit based on the available effects in the material
(e.g. Phong if we have specular, Lambert if we don't have specular but
have diffuse, etc).