Add compare interface for OH::Storage.
Add a compare command in the OH::Storage interface which causes the
transaction to fail if the requested key isn't equal to the given
value. This is a bit more complicated than the bare minimum to support
real locking within the storage layer (e.g. to allow two OH's to
access storage simultaneously and safely resolve conflicts), but its a
good base that results in fewer round trips for individual commands.

Add a few simple tests for the interface and implement it in the
SQLite and Cassandra implementations of OH::Storage.