Fix timestamps for orphan loc updates.
The LocUpdate class was always adjusting timestamps using the server
time offset. However, "orphan" loc updates that are generated from a
removed ProxyObject's properties are already in local time.

This fixes that by making the method that gets the location in local
time virtual and overrides it for the class that wraps
PresenceProperties to provide the LocUpdate interface.