Mark removals due to moving objects between static and dynamic trees as transient.
Update libprox to a version that supports a flag to removeObject that
indicates a removal should be marked as transient. Use this for
removals generated by moving objects between static and dynamic trees
so that these transitions don't cause a small 'outage' for the object
between its removal from one tree (which previously caused the
Permanent flag to be set) and the addition in the other tree.