Cap the number of outstanding location updates per client.
If a client stops responding but remains connected (i.e. the
underlying TCP connection remains but the client stops responding to
message), an unbounded number of location updates could be sent out,
all of which will eventually fail. This eats up memory as it waits for
stream creation to fail and CPU as retries are performed and thousands
of errors are reported.

To fix this, cap the number of outstanding location updates
permitted. This could slow down the time it takes for all updates to
be received (e.g. longer to learn of the steady state if the scene
reaches it), but it ensures we won't have an unbounded number of
messages. An alternative would have been to use a single stream, but
this approach avoids adding unnecessary ordering constraints on the
messages.  Note that the number of messages is the number of
BulkLocationUpdates, so each message may contain many updates to
individual objects.