Tick LocationUpdatePolicy less frequently in BulletPhysicsService.
If you tick LocationUpdatePolicy very quickly in BulletPhysicsService
it generates a lot of updates since moving objects will almost always
need an update. This can result in an overwhelming number of messages,
causing things to get backed up and high latency in actually applying
updates.

This helps with the problem where avatars can get underneath static
terrain. While that can't be fully addressed without additional
changes to BulletPhysicsService, this should help improve the
situation: it increases the average latency of the initial updates
after an object becomes dynamic, but after a few pile up, it actually
decreases the latency since things don't get blocked up and so don't
require retries. It's still possible to use a predicted position which
puts you under the ground, but it's less likely since in these
situations location updates are applied sooner than they are when we
generate a lot more updates.