Add basic Command interfaces, typedefs, and utilities.
The Command classes will allow processing of external commands. To do
so, an optional central Commander will accept commands, track
handlers, and dispatch to them, finally returning results to the
original requestor.

This adds basic classes, typedefs, and interfaces: Command for
flexibly structured commands, CommandKey for uniquely identifying
commands, CommandHandler for processing an individual command, and
Commmander for coordinating CommandHandlers and accepting external
requests. A field is added to Context to make the Commander "globally"
accessible for registrations, but there isn't yet a full
implementation.